Refactor File Converter Test class
